# Conflicts:
#   tests/e2e/testcafe/fixtures/Profile/ViewYourProfile.ts

+++ packages/sdk/src/Sdk.ts
@@ -27,15 +27,18 @@ import notificationCenter from './service/notificationCenter';
 import { SyncService } from './module/sync';
 import { ApiConfig, DBConfig, ISdkConfig } from './types';
 import { AccountService } from './module/account';
-import { UserConfigService } from './module/config';
 import { setGlipToken } from './authenticator/utils';
 import { AccountGlobalConfig } from './module/account/config';
 import { ServiceConfig, ServiceLoader } from './module/serviceLoader';
 import { PhoneParserUtility } from './utils/phoneParser';
 import { configMigrator } from './framework/config';
 import { ACCOUNT_TYPE_ENUM } from './authenticator/constants';
-import { PermissionService, LaunchDarklyController, SplitIOController } from 'sdk/module/permission';
-
+import {
+  PermissionService,
+  LaunchDarklyController,
+  SplitIOController,
+} from 'sdk/module/permission';
+import { jobScheduler } from './framework/utils/jobSchedule';
 
 const LOG_TAG = 'SDK';
 const AM = AccountManager;
@@ -80,7 +83,9 @@ class Sdk {
     );
 
     if (!loginResp || !loginResp.success) {
-      window.indexedDB && window.indexedDB.deleteDatabase('Glip');
+      if (process.env.NODE_ENV !== 'test') {
+        window.indexedDB && window.indexedDB.deleteDatabase('Glip');
+      }
     }
     this._subscribeNotification();
     this._initDataAnalysis();
@@ -101,7 +106,7 @@ class Sdk {
       dbAdapter: dbConfig.adapter,
     });
     Api.init(apiConfig, this.networkManager);
-    await this.daoManager.initDatabase();
+    await this.daoManager.initDatabase(this.clearAllData);
 
     this.permissionService.injectControllers(new LaunchDarklyController());
     this.permissionService.injectControllers(new SplitIOController());
@@ -200,9 +205,6 @@ class Sdk {
     this.networkManager.clearToken();
     this.serviceManager.stopAllServices();
     this.daoManager.deleteDatabase();
-    ServiceLoader.getInstance<UserConfigService>(
-      ServiceConfig.USER_CONFIG_SERVICE,
-    ).clear();
     AccountGlobalConfig.removeUserDictionary();
     this._resetDataAnalysis();
   }
@@ -245,6 +247,16 @@ class Sdk {
   private _resetDataAnalysis() {
     dataAnalysis.reset();
   }
+
+  clearAllData = async () => {
+    await this.daoManager.deleteDatabase();
+    // remove relevant config
+    if (AccountGlobalConfig.getUserDictionary()) {
+      // TODO FIJI-4396
+      this.syncService.userConfig.clearSyncConfigsForDBUpgrade();
+      jobScheduler.userConfig.clearFetchDataConfigs();
+    }
+  };
 }
 
 export default Sdk;
